+++ src/effects/backends/lv2/lv2effectprocessor.cpp
@@ -51,7 +51,7 @@ void LV2EffectProcessor::processChannel(
     }
 
     int j = 0;
-    for (unsigned int i = 0; i < bufferParameters.samplesPerBuffer(); i += 2) {
+    for (SINT i = 0; i < bufferParameters.samplesPerBuffer(); i += 2) {
         m_inputL[j] = pInput[i];
         m_inputR[j] = pInput[i + 1];
         j++;
@@ -66,7 +66,7 @@ void LV2EffectProcessor::processChannel(
     lilv_instance_run(instance, bufferParameters.framesPerBuffer());
 
     j = 0;
-    for (unsigned int i = 0; i < bufferParameters.samplesPerBuffer(); i += 2) {
+    for (SINT i = 0; i < bufferParameters.samplesPerBuffer(); i += 2) {
         pOutput[i] = m_outputL[j];
         pOutput[i + 1] = m_outputR[j];
         j++;