+++ src/effects/backends/lv2/lv2manifest.cpp
@@ -153,9 +153,9 @@ LV2Manifest::LV2Manifest(const LilvPlugin* plug,
 }
 
 LV2Manifest::~LV2Manifest() {
-    delete m_minimum;
-    delete m_maximum;
-    delete m_default;
+    delete[] m_minimum;
+    delete[] m_maximum;
+    delete[] m_default;
 }
 
 QList<int> LV2Manifest::getAudioPortIndices() {
@@ -186,8 +186,9 @@ void LV2Manifest::buildEnumerationOptions(const LilvPort* port,
         const LilvNode* description = lilv_scale_point_get_label(option);
         const LilvNode* value = lilv_scale_point_get_value(option);
         QString strDescription(lilv_node_as_string(description));
-        param->appendStep(qMakePair(
-                strDescription, (double)lilv_node_as_float(value)));
+
+        param->appendStep(qMakePair(strDescription,
+                static_cast<double>(lilv_node_as_float(value))));
     }
 
     if (options != NULL) {